requirements.txt
websocket-client
pillow
pyyaml